Do not suppress 0 values
0 is a valid value for a gauge or a counter (e.g. "number of active sessions"
can easily be zero and that state is distinct from is distinct 1) and zero
samples in a period is legit for a histogram over an interval.

For gauges we should also not reset to zero on each render, but we should
continue to do so for counters.

Suppression of idle metrics should be done using `StatsdBuilder::idle_timeout`.

Fixes #5